server: increase the default for conn_retries
Currently the rpc client uses 0.2 sec sleeps between retries, so with
the old default (3) it was waiting max 0.6 sec before giving up, and
it was not always enough for the spdk target to launch and start
listening the rpc socket.

With the new default (10) it will have about 2 sec. It is not ideal,
as there is still a chance for it to fail (like a testing virtual
environment with highly limited resources), but we probably don't need
a more complex solution right now.

Fixes: https://github.com/ceph/ceph-nvmeof/issues/43
Signed-off-by: Mykola Golub <mykola.golub@clyso.com>