build: add Reviewed-by trailer
This change automates the addition of the `Reviewed-by` trailer. Whenever a user approves a PR, it adds such a trailer to the PR description. If the PR description is used to generate the merge commit summary (which is not the GitHub default), then the `Reviewer-by` trailer becomes automatically appended to the merge commit.

The main con is that the username is the GitHub one, not the one used for the signed-off-by.

Signed-off-by: Ernesto Puerta <37327689+epuertat@users.noreply.github.com>