build: add update-lockfile rule
When a Python package needs to be updated in pyproject.toml, the Python
package manager PDM will check that against the lockfile (pdm.lock) and
if any changes are found it will fail. This ensure builds are
reproducible and consistent.

In order to update, remove or add new packages, the lockfile needs to be
updated. A new Makefile target (update-lockfile) has been added. It runs
on a new container image (nvmeof-builder) and only updates the pdm.lock
file:

```bash
make update-lockfile
git add pdm.lock
```

Signed-off-by: Ernesto Puerta <epuertat@redhat.com>